Time: 0 ms (100%), Space: 8.2 MB (74.19%) - LeetHub

+++ 2469-convert-the-temperature/2469-convert-the-temperature.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     vector<double> convertTemperature(double celsius) {
         return {
-        (double)(celsius + 273.15),
-        (double)(celsius * 1.80 + 32.00)
+        celsius + 273.15,
+        celsius * 1.80 + 32.00
     };
     }
 };
